won't delete profile picture if new one isn't uploaded

+++ welcome.php
@@ -15,7 +15,7 @@
 	$standing=$_POST["year"];
 	$email=htmlspecialchars($_POST["email"], ENT_QUOTES);
 	if(is_numeric($_POST['phn'])){
-					
+		
 		$phone= htmlspecialchars($_REQUEST["phn"], ENT_QUOTES);
 	}
 	else{
@@ -59,19 +59,21 @@
 		//The path you wish to upload the image to
 		$imagePath = "userfolders/$onidid/";
 		$target_file = $imagePath . "profilepic." . basename("$imagetype");
-		if ($_FILES["pic"]["size"] <= 500000) {
-			unlink($imagePath . "profilepic.png");
-			unlink($imagePath . "profilepic.jpg");
-			unlink($imagePath . "profilepic.gif");
-			unlink($imagePath . "profilepic.jpeg");
-			if(move_uploaded_file($imagetemp, $target_file)){
-				chmod ($target_file, 0644);
-				echo("succesfully uploaded image");
-				echo "<br>";
+		if(isset($_POST["pic"])){
+			if ($_FILES["pic"]["size"] <= 500000) {
+				unlink($imagePath . "profilepic.png");
+				unlink($imagePath . "profilepic.jpg");
+				unlink($imagePath . "profilepic.gif");
+				unlink($imagePath . "profilepic.jpeg");
+				if(move_uploaded_file($imagetemp, $target_file)){
+					chmod ($target_file, 0644);
+					echo("succesfully uploaded image");
+					echo "<br>";
+				}
 			}
-		}
-		else{
-			echo "Sorry, your file is too large.";
+			else{
+				echo "Sorry, your file is too large.";
+			}
 		}
 		$myfile=fopen("./userfolders/$onidid/description.txt","w");
 		fwrite($myfile,$desc);
